Move toplevel functions out of Data.SBV
This is in anticipation of introducing the new toplevel Data.SBV.Trans,
which will export more generalized functions than Data.SBV.
